Health/Mana bars, game over scene
Added a health and mana bar to the UI, can increase and decrease them at the moment by pressing 6, 7, 8, and 9. If hitpoints are decreased to 0, it will launch the LoseMenu. LoseMenu is like the main menu, but slightly different to allow for returning to the scene that was lost from.