Revamped resource bars
Revamped the ResourceBar script to be for a class instead of handling the bars in the script. Player health and mana are now stored inside of the PlayerMove script so that the values are attached to the player character. The resource classes for health and mana can now be used and the UI will reflect the values stored in them.